test(FlatteningTest): add flattening type methods test
remove flattening test in MappingTest.kt
